+++ core-datastore/src/main/java/com/hmoa/core_datastore/HShopReview/HShopReviewDataStore.kt
@@ -1,7 +1,6 @@
 package com.hmoa.core_datastore.HShopReview
 
 import ResultResponse
-import com.hmoa.core_model.response.GetMyOrderResponseDto
 import com.hmoa.core_model.response.PagingData
 import com.hmoa.core_model.response.ReviewResponseDto
 import java.io.File
@@ -23,5 +22,4 @@ interface HShopReviewDataStore {
     suspend fun deleteReview(reviewId: Int): ResultResponse<Any>
     suspend fun putReviewLike(reviewId: Int): ResultResponse<Any>
     suspend fun deleteReviewLike(reviewId: Int): ResultResponse<Any>
-    suspend fun getMyOrders(): ResultResponse<List<GetMyOrderResponseDto>>
 }
