+++ core-datastore/src/main/java/com/hmoa/core_datastore/Hshop/HshopRemoteDataStoreImpl.kt
@@ -79,4 +79,13 @@ class HshopRemoteDataStoreImpl @Inject constructor(private val hshopService: Hsh
         }
         return result
     }
+    override suspend fun getMyOrders(): ResultResponse<List<GetMyOrderResponseDto>> {
+        val result = ResultResponse<List<GetMyOrderResponseDto>>()
+        hshopService.getMyOrders().suspendOnError{
+            result.errorMessage = Json.decodeFromString<ErrorMessage>(this.message())
+        }.suspendOnSuccess{
+            result.data = this.data
+        }
+        return result
+    }
 }
